[NO-JIRA] Refactor common src-utils
