+++ src/containers/Admin/Navigation/Container.js
@@ -7,7 +7,7 @@ import { FEATURE_CARD_LAYOUTS } from 'gg-components/FeatureCard';
 import FeatureCard from 'components/common/FeatureCard';
 import STYLES from './admin-navigation.scss';
 import { cssModules } from 'gg-components/helpers/cssModules';
-import { setPostLoginRedirect } from 'utils/storageHelpers';
+import { setPostLoginRedirect } from 'utils/common/storageHelpers';
 
 import Skeleton from './Skeleton';
