+++ src/containers/Admin/Navigation/tests/index.test.js
@@ -3,7 +3,7 @@ import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
 import { initialState as initialAuthenticatorState } from 'containers/common/Authenticator/reducer';
 
-import configureStore from 'utils/redux/configure-store';
+import configureStore from 'utils/common/redux/configure-store';
 
 import AdminNavigationIndex from '../index';
 import AdminNavigation from '../Container';
