+++ src/containers/common/Admin/Users/index.js
@@ -8,8 +8,8 @@ import { load, remove, requestMagicLink, create, update } from './actions';
 import { selectState } from './selectors';
 import { selectState as selectAuthenticatorState } from 'containers/common/Authenticator/selectors';
 import AdminUsers from './Container';
-import injectSaga from 'utils/redux/inject-saga';
-import injectReducer from 'utils/redux/inject-reducer';
+import injectSaga from 'utils/common/redux/inject-saga';
+import injectReducer from 'utils/common/redux/inject-reducer';
 
 import { KEY } from './constants';
 import saga from './saga';
