+++ src/containers/common/Authenticator/index.js
@@ -7,8 +7,8 @@ import { createStructuredSelector } from 'reselect';
 import { loadAuth } from './actions';
 import { selectState } from './selectors';
 import Authenticator from './Container';
-import injectSaga from 'utils/redux/inject-saga';
-import injectReducer from 'utils/redux/inject-reducer';
+import injectSaga from 'utils/common/redux/inject-saga';
+import injectReducer from 'utils/common/redux/inject-reducer';
 
 import { KEY } from './constants';
 import saga from './saga';
