+++ src/containers/common/Consent/index.js
@@ -7,8 +7,8 @@ import { createStructuredSelector } from 'reselect';
 import { consent, deferConsent, resetConsent } from './actions';
 import { selectState } from './selectors';
 import Consent from './Container';
-import injectSaga from 'utils/redux/inject-saga';
-import injectReducer from 'utils/redux/inject-reducer';
+import injectSaga from 'utils/common/redux/inject-saga';
+import injectReducer from 'utils/common/redux/inject-reducer';
 
 import { KEY } from './constants';
 import saga from './saga';
