+++ src/containers/common/Consent/tests/index.test.js
@@ -3,7 +3,7 @@ import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
 import { initialState } from '../reducer';
 
-import configureStore from 'utils/redux/configure-store';
+import configureStore from 'utils/common/redux/configure-store';
 
 import ConsentIndex from '../index';
 import Consent from '../Container';
