+++ src/containers/common/CookiesRequired/tests/index.test.js
@@ -3,7 +3,7 @@ import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
 import { initialState as initialConsentState } from '../../Consent/reducer';
 
-import configureStore from 'utils/redux/configure-store';
+import configureStore from 'utils/common/redux/configure-store';
 
 import CookiesRequiredIndex from '../index';
 import CookiesRequired from '../Container';
