+++ src/containers/common/EmailVerification/tests/index.test.js
@@ -4,7 +4,7 @@ import { Provider } from 'react-redux';
 import { initialState } from '../reducer';
 import { initialState as initialAuthenticatorState } from 'containers/common/Authenticator/reducer';
 
-import configureStore from 'utils/redux/configure-store';
+import configureStore from 'utils/common/redux/configure-store';
 
 import EmailVerificationIndex from '../index';
 import EmailVerification from '../Container';
