+++ src/containers/common/MagicLogin/index.js
@@ -9,8 +9,8 @@ import { selectState } from './selectors';
 import { selectState as selectAuthenticatorState } from 'containers/common/Authenticator/selectors';
 import { selectState as selectConsentState } from '../Consent/selectors';
 import MagicLogin from './Container';
-import injectSaga from 'utils/redux/inject-saga';
-import injectReducer from 'utils/redux/inject-reducer';
+import injectSaga from 'utils/common/redux/inject-saga';
+import injectReducer from 'utils/common/redux/inject-reducer';
 
 import { KEY } from './constants';
 import saga from './saga';
