+++ src/tests/index.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 // import { act } from 'react-dom/test-utils';
 import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
-import configureStore from 'utils/redux/configure-store';
+import configureStore from 'utils/common/redux/configure-store';
 
 import IndexPage from '../pages/index';
