+++ src/utils/common/redux/create-reducer.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 
-import globalReducer from '../../reducer';
+import globalReducer from '../../../reducer';
 
 export default function createReducer(asyncReducers) {
   return combineReducers({
