+++ src/utils/common/redux/tests/check-store.test.js
