+++ src/utils/common/redux/tests/configure-store.test.js
